requirements/dev.txt
-r base.txt

django-debug-toolbar==3.1.1
pylint==2.9.5
